+++ src/cplusplus/leetcode/problem-4.cpp
@@ -65,7 +65,7 @@ private:
     // O(size)
     double findMedianSortedArrays2(vector<int>& nums1, vector<int>& nums2) {
         vector<int> nums;
-        int i = 0, j = 0;
+        size_t i = 0, j = 0;
         for (; i < nums1.size() && j < nums2.size();) {
             if (nums1[i] < nums2[j])
                 nums.emplace_back(nums1[i++]);
